initial data delivery
